restructuration de main
