+++ src/injectData.js
@@ -1,0 +1,27 @@
+import { creationJournal, injecteJournalEcritures } from './journal.js';
+import { creationBalance, injecteBalanceEcritures } from './balance.js';
+import { creationGrandLivre, injecteGrandLivreEcritures } from './grand-livre.js';
+import { creationCompteResultat, injecteCompteResultatEcritures } from './compte-resultat.js';
+import { creationBilan, injecteBilanEcritures } from './bilan.js';
+import { creationInventaire, injecteInventaireEcritures } from './inventaire.js';
+import { creationFEC, injecteFECEcritures } from './fec.js';
+
+export function injectDataIntoPage(jsonData, currentYear, siren) {
+    const mappings = [
+        { id: 'journal-ecritures', create: creationJournal, inject: injecteJournalEcritures },
+        { id: 'balance-ecritures', create: creationBalance, inject: injecteBalanceEcritures },
+        { id: 'grand-livre-ecritures', create: creationGrandLivre, inject: injecteGrandLivreEcritures },
+        { id: 'compte-resultat-ecritures', create: creationCompteResultat, inject: injecteCompteResultatEcritures },
+        { id: 'bilan-ecritures', create: creationBilan, inject: injecteBilanEcritures },
+        { id: 'inventaire-ecritures', create: creationInventaire, inject: injecteInventaireEcritures },
+        { id: 'FEC-ecritures', create: creationFEC, inject: injecteFECEcritures },
+    ];
+
+    mappings.forEach(({ id, create, inject }) => {
+        const element = document.getElementById(id);
+        if (element) {
+            const ecritures = create(jsonData, currentYear, siren);
+            inject(ecritures);
+        }
+    });
+}
