+++ src/loader.js
@@ -1,0 +1,12 @@
+export function showLoader() {
+    document.getElementById('loader').style.display = 'block';
+}
+
+export function hideLoader() {
+    document.getElementById('loader').style.display = 'none';
+}
+
+export function hideErrorMessage() {
+    const errorMessageElement = document.getElementById('error-message');
+    errorMessageElement.style.display = 'none';
+}
