+++ src/setupInfoModal.js
@@ -1,0 +1,19 @@
+export function setupInfoModal() {
+    const infoBtn = document.getElementById('infoBtn');
+    const infoModal = document.getElementById('infoModal');
+    const closeBtn = document.querySelector('.close');
+
+    infoBtn.addEventListener('click', () => {
+        infoModal.style.display = 'block';
+    });
+
+    closeBtn.addEventListener('click', () => {
+        infoModal.style.display = 'none';
+    });
+
+    window.addEventListener('click', (event) => {
+        if (event.target === infoModal) {
+            infoModal.style.display = 'none';
+        }
+    });
+}
